requirements.txt
# Core Application
streamlit==1.48.1

# Vector DB & Embeddings
sentence-transformers==5.1.0
transformers==4.55.2
torch==2.1.2
numpy==1.26.4

# Weaviate Client
weaviate-client==4.16.9

# AWS & LLM
boto3==1.37.10
botocore

# Data Processing & Utilities
langchain==0.3.25
pymupdf==1.26.4
tiktoken==0.9.0
tqdm==4.67.1
pyyaml==6.0.2